refactor: move health check and database metrics monitoring from main to their separate packages (#7)
